+++ visualisation_ok.py
@@ -5,6 +5,7 @@ import glob
 import vtk
 import numpy as np
 import matplotlib.pyplot as plt
+
 # import fenetre_largeur
